refactor: extract tool call state dataclass
